remove some unrequired valueOf

+++ java/src/jmri/jmrit/beantable/oblock/SignalEditFrame.java
@@ -194,7 +194,7 @@ public class SignalEditFrame extends JmriJFrame {
         p1 = new JPanel();
         p1.add(new JLabel(Bundle.getMessage("MakeLabel", Bundle.getMessage("Offset"))));
         lengthSpinner.setModel(
-                new SpinnerNumberModel(Float.valueOf(0f), Float.valueOf(-2000f), Float.valueOf(2000f), Float.valueOf(0.01f)));
+                new SpinnerNumberModel(0f, -2000f, 2000f, 0.01f));
         lengthSpinner.setEditor(new JSpinner.NumberEditor(lengthSpinner, "###0.00"));
         lengthSpinner.setPreferredSize(new JTextField(8).getPreferredSize());
         lengthSpinner.setValue(0f); // reset from possible previous use
